--- app/src/main/java/com/techcourse/handlerMapper/AnnotationHandlerMapping.java
@@ -1,62 +1,0 @@
-package com.techcourse.handlerMapper;
-
-import context.org.springframework.stereotype.Controller;
-import jakarta.servlet.http.HttpServletRequest;
-import org.reflections.Reflections;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import web.org.springframework.web.bind.annotation.RequestMapping;
-import web.org.springframework.web.bind.annotation.RequestMethod;
-import webmvc.org.springframework.web.servlet.mvc.tobe.HandlerExecution;
-import webmvc.org.springframework.web.servlet.mvc.tobe.HandlerKey;
-
-import java.lang.reflect.Method;
-import java.util.HashMap;
-import java.util.Map;
-import java.util.Set;
-
-public class AnnotationHandlerMapping implements HandlerMapping {
-
-    private static final Logger log = LoggerFactory.getLogger(AnnotationHandlerMapping.class);
-
-    private final Object[] basePackage;
-    private final Map<HandlerKey, HandlerExecution> handlerExecutions;
-
-    public AnnotationHandlerMapping(final Object... basePackage) {
-        this.basePackage = basePackage;
-        this.handlerExecutions = new HashMap<>();
-    }
-
-    @Override
-    public void initialize() {
-        Reflections reflections = new Reflections(basePackage);
-        Set<Class<?>> controllers = reflections.getTypesAnnotatedWith(Controller.class);
-        for (Class<?> controller : controllers) {
-            Method[] methods = controller.getDeclaredMethods();
-            for (Method method : methods) {
-                if (method.isAnnotationPresent(RequestMapping.class)) {
-                    RequestMapping annotation = method.getAnnotation(RequestMapping.class);
-                    String url = annotation.value();
-                    for (RequestMethod requestMethod : annotation.method()) {
-                        Object instance = null;
-                        try {
-                            instance = controller.getDeclaredConstructor().newInstance();
-                        } catch (Exception e) {
-                            throw new RuntimeException(e);
-                        }
-                        HandlerKey handlerKey = new HandlerKey(url, requestMethod);
-                        HandlerExecution handlerExecution = new HandlerExecution(instance, method);
-                        handlerExecutions.put(handlerKey, handlerExecution);
-                    }
-                }
-            }
-        }
-        log.info("Initialized AnnotationHandlerMapping!");
-    }
-
-    @Override
-    public Object getHandler(final HttpServletRequest request) {
-        HandlerKey handlerKey = new HandlerKey(request.getRequestURI(), RequestMethod.resolve(request.getMethod()));
-        return handlerExecutions.get(handlerKey);
-    }
-}
