+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/mvc/tobe/handlerMapper/HandlerMapping.java
@@ -1,4 +1,4 @@
-package com.techcourse.handlerMapper;
+package webmvc.org.springframework.web.servlet.mvc.tobe.handlerMapper;
 
 import jakarta.servlet.http.HttpServletRequest;
