+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/mvc/tobe/handlerMapper/ManualHandlerAdapter.java
@@ -1,4 +1,4 @@
-package com.techcourse.handlerMapper;
+package webmvc.org.springframework.web.servlet.mvc.tobe.handlerMapper;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
